Add entropy calculation
Thanks to @julianor's advice, I added an entropy calculation to determinate
correctly the minimum amount of words needed for a secure passphrase,
considering the size of the wordlist and the amount of random numbers
indicated by the user. Also, a warning is shown if the user choses an amount
lower than a secure one.

Change default minimum and maximum for random numbers from (0, 999999) to
(100000, 999999).

Remove version number from man page. Add tip about the insecure amount of words
warning in the README.