Implement Passphrase as a package
Passphrase can be now used as a module inside a script or as a standalone
script. It supports installation via setup.py and makefile.

It is completely self-sufficient, needing no external library, and uses
self-implementation (based on Python's 3.6 Lib/secrets) to get
cryptographically secure random integers from os.urandom().

If numpy is available, it will make use of it for faster entropy calculation.

Fix #1. Fix #3. Fix #4.