Improve password generation
Passwords were generated by randomly choosing items in a single group formed by
the required characters, digits and symbols. The problem with that way
of choosing items is that it doesn't consider the types of the items,
thus getting more chances of choosing a letter than a number, given that
the letters group is bigger than the numbers group.
Now, I create a list with each required type of element (lower case
chars, upper case chars, digits and symbols) and randomly pick one. Then
I pick an item of that group. This way, each type of element has the
same probability of being picked.